+++ NEWS.md
@@ -4,3 +4,5 @@
 
 - `Returns(val)` yields a callable object that returns `val` when called with any
   arguments and keywords.
+
+- `@public` macro to replace the `public` keyword that appears in Julia 1.11.